n2

+++ src/parallel_jobs.jl
@@ -31,7 +31,7 @@ function parallel_null_and_curves(null_size, days_to_event, event, min_threshold
     directions = SharedArray{Float64, 1}((size(expression_mat)[1]))
     threshs = SharedArray{Int, 1}((size(expression_mat)[1]))
     all_pvals_mat = SharedArray{Float64, 2}(size(expression_mat)[1],
-                    floor(size(expression_mat)[1]*(min_threshold+max_threshold)))
+                    Int(floor(size(expression_mat)[1]*(min_threshold+max_threshold))))
 
     # create a channel with the jobs for the null distribution
     println("assigning parallel jobs")
